+++ NEWS.md
@@ -1,5 +1,7 @@
 # teal.modules.general 0.2.7.9000
 
+* Added additional Grouped by subject plot to `tm_missing_data`.
+
 * Issues with no news:
 * Added `max_deg` optional argument to scatterplot to allow user to choose the maximum smoothing degree for trend line.
 * Add new module "tm_outliers" to analyse outlier points.